add formatting layer to getXvsVaultUserInfo
