+++ src/clients/api/queries/getXvsVaultUserInfo/types.ts
@@ -1,0 +1,16 @@
+import BigNumber from 'bignumber.js';
+
+import { XvsVault } from 'types/contracts';
+
+export interface IGetXvsVaultUserInfoInput {
+  xvsVaultContract: XvsVault;
+  tokenAddress: string;
+  poolIndex: number;
+  accountAddress: string;
+}
+
+export interface IGetXvsVaultUserInfoOutput {
+  stakedAmountWei: BigNumber;
+  pendingWithdrawalsTotalAmountWei: BigNumber;
+  rewardDebtAmountWei: BigNumber;
+}
